move transaction components to views/wallet
